requirements.txt
annotated-types==0.7.0
anyio==4.11.0
bcrypt==3.2.2
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
cryptography==46.0.3
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.119.0
filelock==3.20.0
fsspec==2025.9.0
greenlet==3.2.4
h11==0.16.0
huggingface-hub==0.35.3
idna==3.11
Jinja2==3.1.6
lxml==6.0.2
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.5
numpy==2.3.4
packaging==25.0
passlib==1.7.4
pillow==12.0.0
psycopg==3.3.0
psycopg-binary==3.3.0
pyasn1==0.6.1
pycparser==2.23
pydantic==2.12.3
pydantic_core==2.41.4
python-jose==3.5.0
python-multipart==0.0.20
python-pptx==1.0.2
PyYAML==6.0.3
regex==2025.9.18
requests==2.32.5
rsa==4.9.1
safetensors==0.6.2
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.48.0
sympy==1.14.0
tokenizers==0.22.1
torch==2.9.0
tqdm==4.67.1
transformers==4.57.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
xlsxwriter==3.2.9

